requirements.txt
bcrypt==3.2.2
blinker==1.4
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.2
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Dance==6.0.0
Flask-JWT-Extended==4.4.2
Flask-Login==0.6.1
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-SQLAlchemy==2.5.1
Flask-User==1.0.2.2
Flask-WTF==1.0.1
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.1
oauthlib==3.2.0
passlib==1.7.4
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pymongo==4.1.1
pyparsing==3.0.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.39
uritemplate==4.1.1
urllib3==1.26.9
URLObject==2.4.3
Werkzeug==2.1.2
WTForms==3.0.1
